My commit

+++ HostelManagementTeam05/src/main/java/com/avsoft/hostelmanagement/repostiory/BuildingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.avsoft.hostelmanagement.entity.Building;
 
-public interface BuildingRepository extends JpaRepository<Building, Long>{
-
+public interface BuildingRepository extends JpaRepository<Building, Integer> {
+    
 }
